+++ src/TransformerRegistry/TransformerRegistryInterface.php
@@ -23,24 +23,14 @@ interface TransformerRegistryInterface
     public function get(string $id): TransformerInterface;
 
     /**
-     * @param iterable<int,Type|MixedType> $sourceTypes
-     * @param iterable<int,Type|MixedType> $targetTypes
+     * @param iterable<array-key,Type|MixedType> $sourceTypes
+     * @param iterable<array-key,Type|MixedType> $targetTypes
      * @return SearchResult
      */
     public function findBySourceAndTargetTypes(
         iterable $sourceTypes,
         iterable $targetTypes,
     ): SearchResult;
-
-    /**
-     * @param Type|MixedType $sourceType
-     * @param Type|MixedType $targetType
-     * @return iterable<string,TransformerInterface>
-     */
-    public function findBySourceAndTargetType(
-        Type|MixedType $sourceType,
-        Type|MixedType $targetType,
-    ): iterable;
 
     /**
      * @param Type|MixedType $sourceType