+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * fix: PropertyAccessor `UninitializedPropertyException` error now is regarded
   as null.
+* fix: Transformer `SearchResult` was not properly ordered.
 
 ## 0.5.7